+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.3.0] - 2020-12-13
 ### Added
 - option `--debug-cc1101`
 - log MQTT messages before publishing (debug level)
@@ -30,7 +32,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - MQTT client reporting measurements of FT017TH wireless thermo/hygrometers
 
-[Unreleased]: https://github.com/fphammerle/wireless-sensor-mqtt/compare/v0.2.0...HEAD
+[Unreleased]: https://github.com/fphammerle/wireless-sensor-mqtt/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/fphammerle/wireless-sensor-mqtt/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/fphammerle/wireless-sensor-mqtt/compare/v0.2.0...v0.2.0
 [0.1.1]: https://github.com/fphammerle/wireless-sensor-mqtt/compare/v0.1.0...v0.1.1
 [0.1.0]: https://github.com/fphammerle/wireless-sensor-mqtt/releases/tag/v0.1.0